OperationClass Added

+++ SubmoduleClasslibrary/IOperations.cs
@@ -9,6 +9,6 @@ namespace SubmoduleClasslibrary
         int add(int value1, int value2);
         int sub(int value1, int value2);
         int mul(int value1, int value2);
-        int div(int value1, int value2);
+        long div(int value1, int value2);
     }
 }
